pyproject.toml
[project]
name = "scp-stage1-cpt"
version = "0.1.0"
description = "Independent Stage1 CPT pipeline (Hydra + Unsloth + W&B)"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.5.0",
    "transformers>=5.2.0,<=5.3.0",
    "peft>=0.15.0",
    "datasets>=3.0.0",
    "accelerate>=1.3.0",
    "trl>=0.15.0",
    "unsloth",
    "wandb",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "sentencepiece",
    "tiktoken",
    "safetensors",
    "pyyaml",
    "tqdm",
    "numpy",
]

[project.optional-dependencies]
flash = ["flash-attn>=2.7.0", "triton>=3.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
